Fix integer handling in DjvuViewer widget.
git-svn-id: https://svn.master.freemedsoftware.org/freemed/trunk@4283 6507a214-7cff-0310-bd63-e3842a560bc8

+++ ui/gwt/src/org/freemedsoftware/gwt/client/widget/DjvuViewer.java
@@ -165,7 +165,7 @@ public class DjvuViewer extends Composite {
 			}
 			p.NumberOfPages(internalId, new AsyncCallback() {
 				public void onSuccess(Object o) {
-					numberOfPages = (Integer) o;
+					numberOfPages = ((Integer) o).intValue();
 				}
 
 				public void onFailure(Throwable t) {
@@ -183,7 +183,7 @@ public class DjvuViewer extends Composite {
 			}
 			p.NumberOfPages(internalId, new AsyncCallback() {
 				public void onSuccess(Object o) {
-					numberOfPages = (Integer) o;
+					numberOfPages = ((Integer) o).intValue();
 				}
 
 				public void onFailure(Throwable t) {
